pull fennel out into sep file
